Update all projects
Updated TFMs
Updated all packages to latest versions
Set nullable on all projects
Fixed all warnings

+++ JsonParsingBenchmark/Converters/Stj/SearchResultConverter.cs
@@ -171,7 +171,7 @@ namespace JsonParsingBenchmark.Converters.Stj
                         else if (reader.ValueTextEquals(utf8TotalDownloads))
                         {
                             reader.Read();
-                            searchResult.TotalDownloads = reader.GetInt32();
+                            searchResult.TotalDownloads = reader.GetInt64();
                         }
                         else if (reader.ValueTextEquals(utf8Verified))
                         {
